requirements.txt
line-bot-sdk
flask
bs4
pandas
beautifulsoup4